+++ fiddles/three/fiddle-0024-Eclipse/src/metadata.js
@@ -1,0 +1,8 @@
+let metadata = {
+  urls: {
+    github: 'https://github.com/bradyhouse/house/tree/master/fiddles/three/fiddle-0024-Eclipse',
+    moon: {
+      surfaceMaterial: 'resources/images/surfaceMaterial.jpg'
+    }
+  }
+};