+++ fiddles/three/fiddle-0024-Eclipse/src/namespaces.js
@@ -1,0 +1,6 @@
+
+app.toolkit = app.toolkit || {};
+app.toolkit.three = app.toolkit.three || {};
+app.view = app.view || {};
+app.view.milkyway = app.view.milkyway || {};
+